# optimize decode