Speedups:
- eliminating extra passes over data
- `reduce` instead of `reduce_while`
- remove `Vector.add` function call

```
CPU: Intel i7-8550U 1.80GHz
Elixir: 1.14.5-otp-26
Erlang 26.0

optimized (duration: 1.0 s)
benchmark name           iterations   average time
decode small polyline       1000000   1.13 µs/op
decode medium polyline       100000   12.60 µs/op
decode a large polyline        5000   421.46 µs/op

optimized (duration: 60.0 s)
benchmark name           iterations   average time
decode small polyline     100000000   1.19 µs/op
decode medium polyline     10000000   12.18 µs/op
decode a large polyline      500000   396.75 µs/op

vanilla (duration: 1.0 s)
benchmark name           iterations   average time
decode small polyline       1000000   1.66 µs/op
decode medium polyline       100000   23.95 µs/op
decode a large polyline        2000   891.73 µs/op

vanilla (duration: 60.0 s)
benchmark name           iterations   average time
decode small polyline     100000000   1.88 µs/op
decode medium polyline      5000000   26.50 µs/op
decode a large polyline      100000   930.61 µs/op
```