@pkinney 2025-01-26
Released in 1.5.0.  Thank you, @breunigs!